Refactoring MetaClient Exceptions
Refactoring MetaClient Exception. Creating separate exception directory.
