+++ meta-client/src/main/java/org/apache/helix/metaclient/exception/MetaClientInterruptException.java
@@ -1,4 +1,4 @@
-package org.apache.helix.metaclient.constants;
+package org.apache.helix.metaclient.exception;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -19,21 +19,20 @@ package org.apache.helix.metaclient.constants;
  * under the License.
  */
 
-public final class MetaClientTimeoutException extends MetaClientException {
-  public MetaClientTimeoutException() {
+public final class MetaClientInterruptException extends MetaClientException {
+  public MetaClientInterruptException() {
     super();
   }
 
-  public MetaClientTimeoutException(String message, Throwable cause) {
+  public MetaClientInterruptException(String message, Throwable cause) {
     super(message, cause);
   }
 
-  public MetaClientTimeoutException(String message) {
+  public MetaClientInterruptException(String message) {
     super(message);
   }
 
-  public MetaClientTimeoutException(Throwable cause) {
+  public MetaClientInterruptException(Throwable cause) {
     super(cause);
   }
-
 }
